JBR-451

+++ src/main/java/com/jbr/middletier/money/util/TransactionString.java
@@ -1,5 +1,6 @@
 package com.jbr.middletier.money.util;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
@@ -8,7 +9,7 @@ public class TransactionString {
         // Hide the public constructor.
     }
 
-    public static String formattedTransactionString(LocalDate date, double amount) {
+    public static String formattedTransactionString(LocalDate date, BigDecimal amount) {
         return DateTimeFormatter.ofPattern("yyyyMMdd").format(date) + String.format("%.2f",amount);
     }
 }
